requirements.txt
streamlit
langchain
langchain-community
faiss-cpu
python-dotenv
openai
unstructured
tiktoken
numpy
Bio

